Skip compliance operator tests on ROSA (#115)
* Skip compliance operator test on ROSA clusters

The configuration policy cannot create the openshift-compliance
namespace on ROSA clusters because openshift-* namespaces are reserved.
It's possible this situation will come up in other managed environments
so this just checks if the namespace can be created - it doesn't
specifically check for ROSA.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/13759

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix upstream gatekeeper operator namespace

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Skip upstream gatekeeper on ROSA

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update policy URLs to main branch

They were pointing to the master branch, which has been removed. There
is probably some GitHub magic going on which is still making `master`
work, but we shouldn't rely on it.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>