Save canCreateOpenshiftNamespaces result (#117)
In the middle of the gatekeeper test suite, the admission webhook from
gatekeeper can prevent the namespace from being created with an error.
That means some of the tests run, but then the rest are skipped,
including the clean up tests.

By saving the first time it checks if an openshift-* namespace can be
created, the answer can't change before the clean up runs.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/13759

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>