Gatekeeper e2e failures (#126)
* Adjust label selector for gatekeeper operator pod

Depending on the version of the operator (which in turn depends on the
version of the cluster), the gatekeeper pods might be labelled slightly
differently. The new label selector will work for both the "old" and
"new" labels.

Explain relevant details.
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Broaden pattern to check for gatekeeper violation

The message that the gatekeeper gives when denying the request may be
slightly different between versions of gatekeeper. This should work for
more versions.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Stop relying on order of args in gatekeeper

The order of the cmd args in the gatekeeper container may vary between
releases. This will now check that the correct option is present in the
list, without requiring a specific position for it.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update some gatekeeper cleanup logic

Different versions of the operator may or may not automatically remove
the openshift-gatekeeper-system namespace. This will handle both cases,
and ensure the namespace is deleted.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Broaden pattern to check for gatekeeper violation

The message that the gatekeeper gives when denying the request may be
slightly different between versions of gatekeeper. This should work for
more versions.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>