[release-2.3] Backport gatekeeper test updates (#131)
* Backport gatekeeper test updates

The changes to these tests should allow them to function correctly for
both the "old" and "new" versions of the gatekeeper and operator. Those
projects adjusted some details in their deployments and messages.

Refs:
- https://github.com/open-cluster-management/backlog/issues/14957

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update branch to use latest-2.3 images

Since this branch has been cut, we want any tests or deployments from
this branch to use the corresponding images from this branch, NOT from
the main branch.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>