Add a canary test for the IAM policy limitclusteradmin (#139)
* Move the logic of getting the policy compliance state to a shared function

This reduces a lot of code duplication.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Add a canary test for the IAM policy limitclusteradmin

This canary test is added in addition to the existing E2E tests in order to
test that RHBZ#1981127 is resolved. This must be done as a canary
test since it queries for OpenShift group memberships and this is not
supported on other Kubernetes distributions.

Related to:
https://github.com/open-cluster-management/backlog/issues/14166

Signed-off-by: mprahl <mprahl@users.noreply.github.com>