Add metrics tests to run on canaries
The tests in the propagator itself verifies some of these same things,
but the tests here use the service and rbac-proxy set up in the Helm
chart, to verify they are working as expected. One additional test is
to verify that the metrics can't be seen by unauthorized users.

This utilizes the `oc` cli's ability to expose services as a route, so
that we can curl directly to the metrics endpoint. One alternative that
I considered was using a port-forward, but that was difficult to manage.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14992

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>