Pull regex into separate function
This will help document how the regex works, and potentially let it be
used more easily in other metrics tests in the future.

It would have been nice to be able to take multiple labels, but since
golang regex does not have lookaround, we would need to know the order
that the labels will be in.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>