Add metrics tests to run on canaries (#142)
* Add metrics tests to run on canaries

The tests in the propagator itself verifies some of these same things,
but the tests here use the service and rbac-proxy set up in the Helm
chart, to verify they are working as expected. One additional test is
to verify that the metrics can't be seen by unauthorized users.

This utilizes the `oc` cli's ability to expose services as a route, so
that we can curl directly to the metrics endpoint. One alternative that
I considered was using a port-forward, but that was difficult to manage.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14992

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Handle case where there are no policies

If no policies are in the cluster, the metrics endpoint doesn't even
have the HELP comment. So now the test ensures that there is a policy
defined (compliance doesn't matter) that can report some metric.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Use net/http instead of shelling to curl

This will allow for some better timeout handling, and more specific
parsing of the response + response status.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Always try to create the namespace

If the CREATE comes back with an 'already exists' error, then that's
good.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Pull regex into separate function

This will help document how the regex works, and potentially let it be
used more easily in other metrics tests in the future.

It would have been nice to be able to take multiple labels, but since
golang regex does not have lookaround, we would need to know the order
that the labels will be in.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix nitpicks

As mentioned, routeHost is more accurate than routeURL, since it does
not have the protocol.

For `len(authToken) > 0` vs `authToken != ""`, I think I was worried if
the input was a nil string. But you can't have a nil string, so that's
not a problem.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>