Install a specific version of ginkgo for tests
Previously, the `go get` commands were installing the latest versions of
ginkgo and gomega. But the newest version of gomega does not work with
versions of golang less than 1.16. The version of golang installed in
the image is still 1.15, limited by the ubi repos. This now installs a
known good version of ginkgo - matching what is in the go.mod.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>