Fix metrics tests (#145)
* Always specify kubeconfig in oc commands

This makes it more clear where actiona are taking place, and does not
rely on any "default" kubeconfig to have been set.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Enable metrics test

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Move possibly reused functions to common pkg

The metrics functions (with some small adjusments) can be used to test
other metrics endpoints. In particular, the policyreport metrics.

The `ocHub` function seems convenient, especially since the kubeconfig
that it uses is defined in the common package already.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>