Update propagator resource paths
Previously, we had copied resource and CRD yamls into this repo, and
were (mostly) using those for deployment. But those resources can drift,
and it will be better to reference the original locations. With the
changes in the propagator deployment, this is a good time to get it
started.

Where the copied resources are no longer required, I deleted them and
added a README pointing to better locations for them.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/15548

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>