Increase the timeout for the login in the policy generator test
This sets the timeout to five minutes from two minutes. Locally, against
an OCP cluster, this usually takes about 15 seconds, but on some
clusters, this can take longer.

In essence, we are waiting for the cluster authentication operator to
notice the IDP configuration change and modify the oauth-openshift
deployment in order to restart the pods.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>