[release-2.3] Gatekeeper e2e failures (#197)
* Stop relying on order of args in gatekeeper

The order of the cmd args in the gatekeeper container may vary between
releases. This will now check that the correct option is present in the
list, without requiring a specific position for it.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update some gatekeeper cleanup logic

Different versions of the operator may or may not automatically remove
the openshift-gatekeeper-system namespace. This will handle both cases,
and ensure the namespace is deleted.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

Co-authored-by: Justin Kulikauskas <jkulikau@redhat.com>