Use deployment yaml from controllers' repos (#205)
As part of the operator-sdk updates, several controllers have updated
deployment YAMLs for new RBAC and other changes. Previously, the YAML
for deploying the controllers was copied into this repository and
managed separately, but we can just point to the correct path in the
controllers' repos in order to prevent copying and needing to keep them
up to date here.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>