Send travis notifications to a webhook
In order to collect more test summary information, we might want to use
a webhook, as opposed to relying on messages sent through slack. At the
very least, I want to know what information is exactly in the webhook
that we might be able to use.

The endpoint is encrypted, so it doesn't get flooded or anything.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/17431

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>