Grant the subscription-admin user admin access to the test namespace (#279)
The BVT tests started failing recently with the error
"error when retrieving current configuration of..." in the Policy
Generator test. This is due to the user not having read access in
the namespace that the subscription is getting created in.

It's unclear why this is failing now because the created
subscription-admin user has the
open-cluster-management:subscription-admin role, which grants this
access.

This test now no longer relies on the subscription-admin role providing
this access.

See the failing job:
https://app.travis-ci.com/github/stolostron/canary/builds/247478348

Resolves:
https://github.com/stolostron/backlog/issues/20350

Signed-off-by: mprahl <mprahl@users.noreply.github.com>