Add some debug info to try to analyze canary issue
Are we corrupting the hub kubeconfig?  Printing it out to find out.
The one we create is usually less than 900 bytes so I think if we
overwrite the hub kubeconfig with the e2e one it will be obvious from
the trace -- but this won't necessarily help us know why.

Refs:
 - https://github.com/stolostron/backlog/issues/22014

Signed-off-by: Gus Parvin <gparvin@redhat.com>