Use Ordered Describes and After/BeforeAll nodes (#331)
Since fail-fast was turned on for the tests, some of the "clean up" test
nodes are not being run. A better fit for those is "AfterAll" nodes,
which require the Describes to be marked as Ordered.

In addition, some BeforeEach test nodes would fit better as BeforeAll,
since that is now possible.

Note: adding "Ordered" to the Describes largely does not change the
behavior. Unless the `--randomize-all` flag is passed, the It nodes
will be run in the order they are defined anyway.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>