Create and utilize more common functions in e2e
Many of the e2e tests repeatedly create a policy and verify it was
propagated to the managed cluster, so this is now a common function.
This function utilizes the new `patch_decisions` flag (on by default) so
that if the tests are run against a cluster running the PlacementRule
controller, the tests will not "manually" patch the decisions.

Many of the e2e tests needed to check the compliance state of a policy,
so this is another common function. This made many YAML files which had
just been used to compare statuses unnecessary, and they will be cleaned
up in a separate commit.

Many uses of `utils.Kubectl` can be replaced with the (enhanced) `Oc*`
functions, to more consistently use the correct kubeconfig, and more
consistently handle printing and errors.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>