Clean up test policy YAML
As mentioned in the previous commit, many of these were just being used
to compare to for compliant/noncompliant statuses. Those are no longer
necessary with the helper function to check compliance.

Some others needed to be aligned to match the convention required by the
`DoCreatePolicyTest` helper function - namely that PlacementBindings and
PlacementRules have the same names as their corresponding policies, but
are appended with `-pb` and `-plr` respectively.

The gatekeeper YAML previously specified the "default" namespace, but
the namespace should be specified by the command that applies them (and
this allows it to be dynamically adjusted more easily).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>