Add a common cleanup function
Since some policies have multiple templates, this takes the GVR of the
template as a pointer - that way a caller can use `nil` to skip the
template checks, and should then do the correct checks themselves. This
was such a good idea I added it to the Create test func as well, so that
it can check that the policy templates were created.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>