Add common test functions
The new common.Do* functions will make our tests much more concise, and
enforce a slightly more consistent style. By optionally patching the
PlacementRules and configuring that with a common flag, it might be
easier to run these tests on a cluster where that controller is running.

The common.Oc* functions add more consistency to the behavior - the
existing Oc function didn't quite match the propagator's utils.Kubectl
functions. These are also a bit shorter to type. The name is a misnomer,
since the actual client used (oc/kubectl) can be configured via a flag.

This includes one test file converted to use the new functions, as a
demo.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>