Use common functions and prepare for concurrency
This updates the e2e tests to use the new common.Do* functions for many
of their repetitive assertions. It also updates them to label test
containers as `Ordered`, and utilize AfterAll nodes. This makes things
a bit more clear, and allows for cleanup to happen even in the event of
failures. Which prepares it all for running the tests in parallel.

Some YAML files were adjusted for the commond.Do* conventions, and
others were removed because they were just for checking status (which
is now handled differently).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>