E2e common functions (#375)
* Align setup between different suites

This might help move tests between the suites in the future. For the e2e
suite which runs in kind, this requires passing in the desired cluster
namespace ("managed").

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add common test functions

The new common.Do* functions will make our tests much more concise, and
enforce a slightly more consistent style. By optionally patching the
PlacementRules and configuring that with a common flag, it might be
easier to run these tests on a cluster where that controller is running.

The common.Oc* functions add more consistency to the behavior - the
existing Oc function didn't quite match the propagator's utils.Kubectl
functions. These are also a bit shorter to type. The name is a misnomer,
since the actual client used (oc/kubectl) can be configured via a flag.

This includes one test file converted to use the new functions, as a
demo.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Use common functions and prepare for concurrency

This updates the e2e tests to use the new common.Do* functions for many
of their repetitive assertions. It also updates them to label test
containers as `Ordered`, and utilize AfterAll nodes. This makes things
a bit more clear, and allows for cleanup to happen even in the event of
failures. Which prepares it all for running the tests in parallel.

Some YAML files were adjusted for the commond.Do* conventions, and
others were removed because they were just for checking status (which
is now handled differently).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Allow variations in k8s client

Installing things in Prow is tricky, so those tests can just use what
already works: kubectl. Other tests will now install both clients, and
can be configured to use either.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>