Add qualified policy type name and pin gatekeeper operator version (#386)
* Add qualified policy type name

Kyverno policies are also just called "policy", so the tests need to be
a little more specific when dealing with open-cluster-management.io
policies through kubectl commands.

Refs:
 - https://github.com/stolostron/backlog/issues/23005

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Pin version of community gatekeeper operator

The latest tag was just updated, and is not guaranteed to work.

Refs:
 - https://github.com/stolostron/backlog/issues/23005

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Increase timeout for kyverno deployment

The deployment gets created when the subscription installs the kyverno
helm chart.  Sometimes this can take a while for the helm install to
start so the default timeout is really a bit short.  Increasing it to
reduce the concern here.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>