Improve policy_governance_info metric RBAC tests
In newer versions of Kubernetes, tokens are not automatically created
for all service accounts. Therefore, the tests might need to manually
create the tokens for the service accounts it uses.

There is also a new test for users that should not have access to the
endpoint, whereas previously we only tested unauthenticated access.

Refs:
 - https://github.com/stolostron/backlog/issues/24075

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>