Add tests for pruneObjectBehavior feature
Tests the general functionality of this new feature on KinD.

These tests are in a new (experimental) format that would allow them to
be run in our usual KinD tests (under e2e) as well as in our larger set
of tests that run against real clusters (under integration). This
format would also potentially allow us to turn tests on and off
separately from the ginkgo filters, which could be helpful because the
filters cause tests to be *skipped*, which impacts test execution stats.

Refs:
 - https://github.com/stolostron/backlog/issues/24440

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>