Add reusable workflow for running KinD tests
Previously, we added actions to many component repositories to run these
tests on their own PRs, helping to ensure things continued to work well
together. This brings that work into this repository, so that those
repositories do not need to maintain the workflow individually.

It also includes new enhancements to improve the experience on GitHub,
for example saving debug information for download during a failure, and
saving test reports for possible future use. These are accessible from
the summary page of the action run.

Finally, it adds a local call of the workflow on pushes and PRs, which
could replace some other tool's invocation of the tests.

Refs:
 - https://github.com/stolostron/backlog/issues/22914

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>