Trying a new idea around validating the compliance operator status
We want the scan to be Running and ideally get to aggregating but that
seems to time out a lot.  I think if it stays running we can still
consider the test OK -- but it could also transition to aggregating or
done so this PR will let the final state pass in any of these cases.

Refs:
 - https://github.com/stolostron/backlog/issues/25631

Signed-off-by: Gus Parvin <gparvin@redhat.com>