Simplify the Gatekeeper tests
Based on cf34fcd1bf63347c1812c7f435cf21f25afd9019

Testing the Gatekeeper functionality is out of scope for the BVT/SVT/E2E
tests since that is limited to the ACM Governance functionality. This
should be handled by the Gatekeeper and Gatekeeper operator tests
instead.

This was prompted by these tests often taking longer than the max
timeout of 10 minutes that is set, and thus leading to flaky BVT and SVT
results.

Refs:
 - https://github.com/stolostron/backlog/issues/26656

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>