Polish ConfigurationPolicy pruning test
It seems that if the ConfigurationPolicy hadn't been updated with a
status, it was vulnerable to deletion, so this adds a check for its
to be up-to-date.

Additionally, there was no timeout for the ConfigMap with a finalizer,
so the command would continue to wait indefinitely. This adds a timeout
as well as a patch to remove the finalizer.

Signed-off-by: Dale Haiducek <19750917+dhaiducek@users.noreply.github.com>