Test ManagedClusterLabels in hub templates
One test is added to the BVT, one is SVT-only. I was planning on having
an additional test to check that if a label is queried that doesn't
exist, then the policy would be non-compliant (I thought that such an
error would appear as a template-error), but the Go stdlib `template`
inserts `<no value>` in that situation. That might not be ideal, but
it's not something to fix at this moment.

Refs:
 - https://issues.redhat.com/browse/ACM-4177

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>