Add debug messages for some operator policy fails
The new `RegisterDebugMessage` is based off of a pattern that is working
well in the config-policy-controller tests in order to only print
certain debug information when the test fails. This is especially useful
in `Eventually` assertions.

This updates several tests to use this new function which will hopefully
help in debugging some inconsistent tests.

Refs:
 - https://github.com/stolostron/backlog/issues/27556
 - https://github.com/stolostron/backlog/issues/27558

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>